requirements.txt
# Local gRPC stub package (editable install)
-e ../waw-contracts

#sqlcipher3==0.5.4
requests
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
grpcio==1.64.0
grpcio-tools==1.64.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlcipher3-binary==0.5.4
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
